refactor(templates): improve codebase quality and modernize patterns
Modernize lint configurations and add CI resilience for clippy phantom-cache
replay. Specifically:

- Add `#![allow(clippy::unwrap_used, clippy::expect_used, clippy::panic, missing_docs)]`
  headers to every `#[cfg(test)] mod tests` block and benchmark file so
  test-grade diagnostics aren't blocked by `-D warnings` in cargo clippy.
- Document the rationale in `[workspace.lints.clippy]` of Cargo.toml.
- Add a `find target/.fingerprint -maxdepth 1 -name 'rust-2026-template-*'
  -print -delete` step before the Cargo clippy invocation in CI, since
  `cargo clean -p` does not remove the `.fingerprint/` directory that
  Rust 1.66+ uses for stale diagnostic replay.
- Set `cache-on-failure: false` on the `Swatinem/rust-cache` invocation
  in `.github/actions/setup-rust/action.yml` so failed clippy runs do not
  re-upload poisoned target/ artifacts.
- Add `examples/roast-scorer.rs` as a real, properly-lived Rust example
  that demonstrates `Result`-returning `main` with `?`-propagation under
  the workspace's `-D warnings` policy.

Conflicting with origin/main's pre-existing roast-scorer example was
resolved by accepting HEAD's clippy-clean version.
