+++ .github/actions/setup-rust/action.yml
@@ -26,4 +26,7 @@ runs:
 
     - uses: Swatinem/rust-cache@e18b497796c12c097a38f9edb9d0641fb99eee32
       with:
-        cache-on-failure: true
+        # Don't save the cache when the workflow fails — otherwise a Clippy
+        # failure leaves phantom `.fingerprint/` metadata in `target/` that
+        # gets re-uploaded, replaying stale diagnostics on the next PR.
+        cache-on-failure: false
