+++ benchmarks/Cargo.toml
@@ -10,9 +10,11 @@ workspace = true
 [dependencies]
 rust-2026-template = { path = ".." }
 sample-app = { path = "../crates/sample-app" }
+example-crate = { path = "../crates/example-crate" }
 
 [dev-dependencies]
 criterion = { workspace = true }
+tempfile = "3"
 
 [[bench]]
 name = "end_to_end"
