+++ crates/actor-runtime-template/src/lib.rs
@@ -24,6 +24,8 @@
 //! - Supervision trees with restart policies
 //! - Graceful shutdown
 
+#![forbid(unsafe_code)]
+
 pub mod actor;
 pub mod supervisor;
 
@@ -140,6 +142,8 @@ pub trait ActorState: Send + Sync + Clone + 'static {
 
 #[cfg(test)]
 mod tests {
+    #![allow(clippy::unwrap_used, clippy::expect_used, clippy::panic, missing_docs)]
+
     use super::*;
 
     #[test]
