+++ crates/checkpoint-template/src/storage.rs
@@ -142,6 +142,8 @@ impl FileStorage {
 
 #[cfg(test)]
 mod tests {
+    #![allow(clippy::unwrap_used, clippy::expect_used, clippy::panic, missing_docs)]
+
     use super::*;
     use tempfile::TempDir;
