+++ crates/mcp-server-template/Cargo.toml
@@ -16,7 +16,6 @@ workspace = true
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
 thiserror = { workspace = true }
-async-trait = { workspace = true }
 rmcp = { workspace = true, features = ["server", "macros"] }
 tokio = { workspace = true, features = ["sync"] }
 tracing = { workspace = true }
