+++ crates/sample-app/Cargo.toml
@@ -15,10 +15,6 @@ include = ["/src", "README.md", "LICENSE"]
 [lints]
 workspace = true
 
-[lib]
-name = "sample_app"
-path = "src/main.rs"
-
 [[bin]]
 name = "sample-app"
 path = "src/main.rs"
@@ -32,3 +28,4 @@ tracing-subscriber.workspace = true
 clap.workspace = true
 
 [dev-dependencies]
+tempfile = "3"
