+++ examples/hello_world/src/main.rs
@@ -1,6 +1,9 @@
+//! Hello World example demonstrating the example-crate.
+
 use anyhow::Result;
 use example_crate::greet;
 
+/// Main entry point.
 fn main() -> Result<()> {
     let message = greet("World");
     println!("{message}");
