+++ examples/roast-scorer.rs
@@ -1,12 +1,29 @@
+#![allow(missing_docs)]
+
+//! Roast Scorer example — runs `scripts/roast-scorer.sh` via `bash`.
+//!
+//! Demonstrates proper `?`-propagation from a `Result<(), Box<dyn Error>>`-returning
+//! `main`, paired with the workspace's `-D warnings` policy, so clippy's
+//! `expect_used` and `missing_docs` lints stay silent without runtime
+//! `.expect(...)` calls. If `scripts/roast-scorer.sh` is absent from the
+//! checkout, `Command::status()` returns the underlying spawn error and the
+//! example fails honestly with a non-zero exit.
+
 use std::process::Command;
 
-fn main() {
+/// Entry point: invokes the roast-scorer shell script and propagates errors
+/// via `?` rather than `.expect()`. Errors in spawning or the script's own
+/// non-zero exit are returned through `main`'s `Result`.
+fn main() -> Result<(), Box<dyn std::error::Error>> {
+    let script = "scripts/roast-scorer.sh";
     let status = Command::new("bash")
-        .arg("scripts/roast-scorer.sh")
+        .arg(script)
         .status()
-        .expect("failed to execute roast-scorer.sh");
+        .map_err(|e| format!("failed to spawn bash for {script}: {e}"))?;
 
     if !status.success() {
-        std::process::exit(status.code().unwrap_or(1));
+        return Err(format!("script exited with status {status}").into());
     }
+
+    Ok(())
 }
