+++ fuzz/fuzz_targets/fuzz_sanitize.rs
@@ -3,6 +3,6 @@ use libfuzzer_sys::fuzz_target;
 
 fuzz_target!(|data: &[u8]| {
     if let Ok(s) = std::str::from_utf8(data) {
-        let _ = rust_2026_template::add(0, s.len() as u64);
+        let _ = sample_app::sanitize_str(s);
     }
 });
